Unpin a bunch of things
Let shade install its deps. We control that anyway. Also, use non-pinned
ansible. Its pinned in ursula because they're on a stable release and
didn't wnat to verify just right now.

If there are shade problems, or problems with ansible shade modules we
can fix them quickly, and if that's not quick enough, we CAN track them
here locally too.

Finally, we do not need tox or pep8 in an install.