Don't recursively clone superproject
This is slow and painful on both the localhost and servers. In all cases
from CI we are going to know the branch name - it's almost always going
to be deploymaybe and hopefully the githook will give us an actual ref.
We can pass that through to ansible as a command line argument.

Clone the superproject but only clone the projects that we actually need
from it as we need them.