Use keystone-manage bootstrap to setup admin user
The V2 default domain is no longer automatically created. This will
probbaly be fixed upstream but the intended solution is to use bootstrap
to create your admin user, project and role and this will create a
default domain.

In future this will allow us to not use the admin_token at all, but for
now lets just get this passing.