Just clone recursively at the top
In clone-source we already pre-prune the superrepo for what we need.
That means we will, over the course of running arrrsula, update every
submodule in the superrepo. If we do it all at once at the top, we
reduce the chances of race conditions having done something strange with
the git refs up on the server.

Change-Id: Ib024891e267ed09fde2c78197d14757f9961ce61