Use is defined for when: provider_interface
For some reason, after we run a role conditionally with "when:
provider_interface", the variable provider_interface variable is set to
a value which ansible cannot evaluate. Any attempts to print the
variable after this point result in a 'conditional error'. If we use
when: provider_interface is defined instead, this does not occur.

Change-Id: I6ff47f01e6b56a12ba1ec866e77d6a6ee86f10b0