Work using resolvconf updater
When resolvconf is installed on the base images and you install dnsmasq
it will take over the dns route. Instead of guessing always ensure
resolvconf is installed and setup dns via the application so later
when dnsmasq bounces resolvconf we end up with a real resolv
configuration.

As part of this we move the resolvconf role into the common role rather
than be a dependency. It is never otherwise consumed and makes more
sense in common.

Change-Id: Idee677d1d0b454f176bd4543094048815b951a35