Fetch git tags with superrepo
So it appears the problem with submodules is that the reference is not
public and available at the time of the clone. To fix this we'll start
tagging deploymaybe releases. To make sure the tag we want is available
at submodule update time fetch all the tags of the superproject.

Change-Id: Ie06bc4aaf3dc5086021840f9f10a394d9cba7b01