Only setup bridges in prereboot
We don't need to setup monitoring or anything else, just the bridges so
that the prereboot tasks that setup /e/n/i can do so before we reboot
with the new network configuration.

This fixes the issue where prereboot fails because ovsdb-server isn't
running. That may also point to another bug which is that we aren't
starting it in time.

Change-Id: I787c51415a0cd334110597346c9cdc8deb3c00e2