Don't cache catalog entries
There doesn't seem to be a catalog in the token authentication responses
immediately after creating the endpoints which i'm pretty sure is a
caching issue.

This is also the config used in devstack that we should look at fixing
upstream as catalogs are one of the main things to want to cache.

Change-Id: I0954e67e3c5ed9eb61838b16a5d3e1ab4fc3034e