Add prereboot tag to apt repo tasks
The prereboot tag includes a task that changes the apt
repository for general packages, but does not depend on
apt-repos which enables installation of unauthenticated
packages. Thus, the kernel-tuning tasks that run during
the prereboot phase fail during package installation.

Include apt-repos and update the package index during
the prereboot phase.

Change-Id: Ie91e65565a2851ee87d82e01829a4d056311e213