Use a tarball to push git repos onto servers.
This goes very slowly for small deployments, as creating a tarball of
the entire repo is slow. However, at scale this is quite a bit faster as
long as the arrrsula node can keep up with pushing the file out.  Also,
in CI, this will lead to pushing what is checked out by zuul-cloner,
rather than needing to translate that into something we can inject into
the settings.

Change-Id: I74bffed1009ca95539ed7f8d2c275aa90ab01f10