Rename ml2_plugin.ini to ml2_conf.ini
The neutron community file is named: ml2_conf.ini.  I looked back all
the way back to the Havana release and I can not find any case in which the
community was using: ml2_plugin.ini

The file: ml2_plugin_dataplane.ini is an ursula specific file and does
not exist in the community, but I renamed it as well to be consistent.

Change-Id: I030e0af01950fedf1560b113defe3e94d2e841eb