Fix bug in bounce network interfaces handler
This was supposed to only bounce interfaces marked as "safe for
reconfiguration" with 'allow-reconfiguration'. But the -a argument
actually does all of the ones marked as -a auto, which is usually the
management interface Ansible uses, and so, if the ifdown doesn't happen
quick enough, we'll get hung and close connection before ifup happens.

Change-Id: I43c91dcbac626337792004a5ca93bb7ea7b216f3