Added newline to the end of br-provider.cfg
During our testing we found that the default route was not being
assigned as a result of br-provider.cfg not having a newline at the end
of the file.

Also had to change /etc/network/interfaces.d/bond1.cfg to use manual
instead of static, since no IP address is assigned to it.

Change-Id: Ica562b4428642bb0086bf04dbf9e834f8f46dfb3
Signed-off-by: Matt Langbehn <mlangbe@us.ibm.com>