force newline in br-provider.cfg
This will add a newline variable set to a \n delimiter to be used in
br-provider.cfg. If this newline is not forced, it does not get
correctly interpreted by ansible, and the br-provider config that is
generated does not work.

add newline to bond1.cfg as well for consistency

Change-Id: I2f51fd6a63ac2a2e6e15c230461c0121f8aff24b