Fix for mysql 'too many connections'.
This change reduces the number of Keystone proccesses per port to 10 and
increases the max_connections for Percona to 2k. Originally we were using
ansible_processor_vcpus for both the number of Keystone procs and threads.
This meant with our 3 controller nodes, each with 56 cpus, and each running
the public and admin ports, we had 3 * 56 * 56 * 2 = 18,816 connections to
Percona while our max_connections only allowed 1k.

With a previous change we dropped to a fixed 10 threads for Keystone. Now
this change drops Keystone's number of procs to 10 as well. This means we
will now have 3 * 10 * 10 * 2 = 600 connections to Percona from Keystone.
This change also bumps the max_connections for Percona up to 2k so that
we have additional room for other services that need to connect.

Change-Id: I0124b4afb9c5a0f59b0155199e621da5662e1595
Signed-off-by: Matt Mulsow <mamulsow@us.ibm.com>