Remove extraneous OVSDB server on DB node
Previously it was observed that an additional OVSDB server was
being started on the DB node in addition to the NB and SB ones.
This third DB was causing problems as it sometimes would use
either the NB or SB port. This patch turns off the service
openvswitch-switch which removes the extra DB.

Change-Id: I084dadf532a60b77be31c5729d82acf5bbf33b5c