Remove ssh check after kernel upgrade
There appears to be a bug in ansible that causes wait_for tasks
to take the full timeout period under unknown circumstances
even when the condition they are checking for has been met well
before the timeout has been reached.  See

http://logs.cloud-ci.ibmcis.com/29/1929/2/experimental/ci-tempest/6f26a51/console.html.gz

for an example where the timeout has been set to 20 minutes and the
task takes 40 minutes because 5 tasks can be run in parallel and there
are six hosts.  Removing the check and relying on the initial check
during the subsequent run (when the kernel tags are skipped) does not
exhibit this behavior.  See:

http://logs.cloud-ci.ibmcis.com/96/1996/6/experimental/ci-tempest/0c8d415/console.html

for an example.  This change works for the use case of running arrrsula
first with the kernel and prereboot tags and skipping the tasks marked
with the kernel tag during a subsquent run.  Running arrrsula for the first
time against an environment without any specified or skipped tags would
fail because there is no delay between when the systems are rebooted after
the kernel upgrade and subsequent steps.

Change-Id: I353241d387b5ebc78d17d467fb9bd782b98d16ed