Increased HAProxy maxconn values
We have been seeing HAProxy returning 503 responses when we do
any sort of concurrency testing in the PSR and IB environments.
This seems to be caused by a large number of connections sitting
in the queue and eventually timing out. Increasing this value
should help us handle more of these requests in parallel.

Change-Id: Ic47edf44be249308d3ae5a6d59073ed324cab45b
Signed-off-by: Matt Mulsow <mamulsow@us.ibm.com>